+++ solvers/rl_solver.py
@@ -1,6 +1,6 @@
 """Solver for Reinforcement Learning (Q-Learning)"""
 
-__version__ = "0.4.1"
+__version__ = "0.5.0"
 __author__ = "ariadie@gmail.com"
 __date__ = "2025-12-24"