+++ solvers/sa_solver.py
@@ -1,13 +1,13 @@
 """Solver for Simulated Annealing"""
 
-__version__ = "0.4.1"
+__version__ = "0.5.0"
 __author__ = "ariadie@gmail.com"
 __date__ = "2025-12-24"
 
 import random
 import math
 from typing import List, Any
-from problem import Problem
+from problems import Problem
 from logger import Logger
 
 class SASolver: